Revert "Adjust Intel SCU and Watchdog drivers for Edison"
This reverts commit 0b247ec2a0e8a2e5509064a222ba1cf9cac977bd.

Reviewed-on: https://android.intel.com/424750

+++ arch/x86/include/asm/intel_scu_ipc.h
@@ -64,8 +64,8 @@ int intel_scu_ipc_check_status(void);
 
 /* Issue commands to the SCU with or without data */
 int intel_scu_ipc_simple_command(int cmd, int sub);
-int intel_scu_ipc_command(int cmd, int sub, u32 *in, int inlen,
-				u32 *out, int outlen);
+int intel_scu_ipc_command(u32 cmd, u32 sub, u8 *in, u32 inlen,
+				u32 *out, u32 outlen);
 /* I2C control api */
 int intel_scu_ipc_i2c_cntrl(u32 addr, u32 *data);
